requirements.txt
asgiref==3.5.2
Django==4.1.3
factory-boy==3.2.1
Faker==15.3.2
mysqlclient==2.1.1
python-dateutil==2.8.2
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.3
tzdata==2022.6
